requirements.txt
# Requirements for plz-api

flask
python-dotenv
psycopg2-binary
argon2-cffi